# Conflicts:
#	backend/de.metas.adempiere.adempiere/base/src/main/java-legacy/org/adempiere/ad/dao/impl/TypedSqlQuery.java
#	backend/de.metas.async/src/main/java/de/metas/async/processor/impl/WorkpackageProcessorTask.java
#	backend/de.metas.async/src/test/java/de/metas/async/processor/impl/WorkpackageProcessorTaskTest.java
#	backend/de.metas.business.rest-api-impl/src/main/java/de/metas/rest_api/v2/ordercandidates/impl/OrderCandidateRestControllerService.java
#	backend/de.metas.cucumber/src/test/java/de/metas/cucumber/stepdefs/invoice/C_Invoice_StepDef.java
#	backend/de.metas.cucumber/src/test/resources/de/metas/cucumber/features/createProductionOrder.feature
#	backend/de.metas.cucumber/src/test/resources/de/metas/cucumber/features/materialDispo/production.feature
#	backend/de.metas.cucumber/src/test/resources/de/metas/cucumber/features/materialDispo/supplyRequiredQtyViaPurchase.feature
#	backend/de.metas.cucumber/src/test/resources/de/metas/cucumber/features/shipment/reversedShipment.feature
#	backend/de.metas.swat/de.metas.swat.base/src/main/java/de/metas/invoice/InvoiceService.java
#	backend/de.metas.swat/de.metas.swat.base/src/main/java/de/metas/invoicecandidate/api/IInvoiceCandInvalidUpdater.java
#	backend/de.metas.swat/de.metas.swat.base/src/main/java/de/metas/invoicecandidate/api/IInvoiceCandidateEnqueuer.java
#	backend/de.metas.swat/de.metas.swat.base/src/main/java/de/metas/invoicecandidate/api/impl/InvoiceCandidateEnqueuer.java
#	backend/de.metas.swat/de.metas.swat.base/src/main/java/de/metas/invoicecandidate/api/impl/InvoiceCandidatesChangesChecker.java
#	backend/de.metas.swat/de.metas.swat.base/src/main/java/de/metas/invoicecandidate/async/spi/impl/InvoiceCandWorkpackageProcessor.java
#	backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/menu/datatypes/json/JSONMenuNode.java

+++ backend/de.metas.shipper.gateway.derkurier/src/main/java/de/metas/shipper/gateway/derkurier/DerKurierDeliveryOrderRepository.java
@@ -37,7 +37,7 @@ import lombok.NonNull;
 import org.adempiere.ad.dao.IQueryBL;
 import org.adempiere.ad.dao.IQueryBuilder;
 import org.adempiere.model.InterfaceWrapperHelper;
-import org.adempiere.util.lang.IPair;
+import de.metas.common.util.pair.IPair;
 import org.adempiere.util.lang.ITableRecordReference;
 import org.adempiere.util.lang.impl.TableRecordReference;
 import org.compiere.model.I_C_Country;
@@ -250,7 +250,7 @@ public class DerKurierDeliveryOrderRepository implements DeliveryOrderRepository
 					.build();
 			deliveryPositionBuilder.packageDimensions(packageDimensions);
 		}
-		deliveryPositionBuilder.grossWeightKg(lineRecord.getDK_ParcelWeight().intValue());
+		deliveryPositionBuilder.grossWeightKg(lineRecord.getDK_ParcelWeight());
 
 		final DerKurierDeliveryData derKurierDeliveryData = DerKurierDeliveryData.builder()
 				.station(assertSameAsPreviousValue(COLUMNNAME_DK_Consignee_DesiredStation, lineRecord, previousLineRecord))
@@ -411,7 +411,7 @@ public class DerKurierDeliveryOrderRepository implements DeliveryOrderRepository
 				lineRecord.setDK_ParcelLength(BigDecimal.valueOf(packageDimensions.getLengthInCM()));
 				lineRecord.setDK_ParcelWidth(BigDecimal.valueOf(packageDimensions.getWidthInCM()));
 			}
-			lineRecord.setDK_ParcelWeight(BigDecimal.valueOf(deliveryPosition.getGrossWeightKg()));
+			lineRecord.setDK_ParcelWeight(deliveryPosition.getGrossWeightKg());
 
 			lineRecord.setDK_ParcelNumber(derKurierDeliveryData.getParcelNumber());
